requirements.txt
discord.py
tabulate
table2ascii
pytz
